# src/auslib/web/admin/swagger/api.yml
swagger: '2.0'
info:
  version: '1.0'
  title: "Balrog's Admin API"
  termsOfService: "https://wiki.mozilla.org/Balrog"
  license:
    name: "Mozilla Public License 2.0"
    url: "https://www.mozilla.org/en-US/MPL/2.0/"
  contact:
    name: "Mozilla Release Engineering"
    url: "https://github.com/mozilla-releng/balrog"
    email: "release-engineering@lists.mozilla.org"
  description: >
    **Balrog** is the software that runs the server side component of the update
    system used by Firefox and other Mozilla products. It is the successor to *AUS*
    (Application Update Service), which did not scale to our current needs nor allow
    us to adapt to more recent business requirements. Balrog helps us ship updates faster
    and with much more flexibility than we’ve had in the past.

# host and scheme can be omitted for a more dynamic association.
# In this case, the host and scheme used to serve the API documentation
# will be used for API calls. For example, if Swagger UI-based documentation is
# hosted at https://api.example.com/apidocs/index.html, “try it out” API calls
# will be directed to https://api.example.com.

#schemes:
#  - http
#  - https
# basePath: "/api"
# host: "localhost:8080"
# Swagger-UI bug : fix TODO week 13
...
parameters:
...
  dataVersionParam:
    name: data_version
    in: query
    description: data version of object.
    type: integer
    format: int32
    minimum: 1
    required: true
...
  change_id_body_param:
    name: object_change_id
    in: body
    description: change_id of the Rule/Release/Scheduled_Change to revert back to
    required: true
    schema:
      type: object
      required:
        - change_id
      properties:
        change_id:
          type: integer
          description: change_id of the Rule/Release/Scheduled_Change revision to revert to
          minimum: 0
          example: 85142
...
responses:
  updateExistingObject:
    description: "successfully updated the existing object and incremented its data_version"
    schema:
      type: object
      required:
        - new_data_version
      properties:
        new_data_version:
          description: "data version of the updated object. Should be greater than 1"
          type: integer
          format: int32
          minimum: 2
          example: 2
...
  modifiedDataVersion:
    description: "Created or Updated exisiting object."
    schema:
      type: object
      required:
        - new_data_version
      properties:
        new_data_version:
          description: "data version of the updated or new object"
          type: integer
          format: int32
          minimum: 1
          example: 2
...
paths:
...
  /releases/{release}:
    get:
      summary: "Returns the 'data' portion of the named Release, which is a JSON Object."
      description: >
        Returns the 'data' portion of the named Release, which is a JSON Object. If 'pretty' is present in the query string and set to true, it will be pretty formatted. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#releases-release)
      tags:
        - Releases
      operationId: auslib.web.common.releases.get_release
      consumes: []
      produces:
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#releases-release"
        description: "Returns the 'data' portion of the named Release, which is a JSON Object."
      parameters:
        - $ref: '#/parameters/releaseParam'
        - $ref: '#/parameters/pretty'
      responses:
        '200':
          $ref: '#/responses/releaseDataResponse'
        '404':
          $ref: '#/responses/resourceNotFound'

    put:
      summary: Overwrites the named Release with the data given.
      description: >
        Overwrites the named Release with the data given. The 'blob' field is completely overridden with the new one, not updated. If the Release does not exist, it is created. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#put)
      tags:
        - Releases
      operationId: auslib.web.admin.views.releases.ensure
      consumes:
        - application/json
      produces:
        - text/html
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#put"
        description: "Overwrites the named Release with the data given."
      parameters:
        - $ref: '#/parameters/releaseParam'
        - name: release_body
          in: body
          description: Release object data
          required: true
          schema:
            $ref: '#/definitions/ReleaseData'
      responses:
        '200':
          $ref: '#/responses/updateExistingObject'
        '201':
          description: successfully created a new release
        '400':
          $ref: '#/responses/invalidFormData'
        '401':
          $ref: '#/responses/unauthenticatedUser'
        '403':
          $ref: '#/responses/unauthorizedUser'

    post:
      summary: Updates the named Release with the data given.
      description: >
        Updates the named Release with the data given. The 'blob' field is updated with the new one instead of being completely overridden. If the Release does not exist, it is created. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id11)
      tags:
        - Releases
      operationId: auslib.web.admin.views.releases.update
      consumes:
        - application/json
      produces:
        - text/html
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id11"
        description: "Updates the named Release with the data given."
      parameters:
        - $ref: '#/parameters/releaseParam'
        - name: partial_release_body
          in: body
          description: Release object data
          required: true
          schema:
            $ref: '#/definitions/PartialRelease'
      responses:
        '200':
          $ref: '#/responses/modifiedDataVersion'
        '201':
          $ref: '#/responses/modifiedDataVersion'
        '400':
          $ref: '#/responses/invalidFormData'
        '401':
          $ref: '#/responses/unauthenticatedUser'
        '403':
          $ref: '#/responses/unauthorizedUser'

    delete:
      summary: Deletes the named Release
      description: >
        Deletes the named Release. The 'data_version' parameter must be provided. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id12).
      tags:
        - Releases
      operationId: auslib.web.admin.views.releases.delete
      consumes: []
      produces:
        - application/json
        - text/html
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id12"
        description: Deletes the named Release
      parameters:
        - $ref: '#/parameters/releaseParam'
        - $ref: '#/parameters/dataVersionParam'
      responses:
        '200':
          description: Successfully deleted the release
        '401':
          $ref: '#/responses/unauthenticatedUser'
        '403':
          $ref: '#/responses/unauthorizedUser'
        '404':
          $ref: '#/responses/resourceNotFound'
...
  /releases/{release}/builds/{platform}/{locale}:
    get:
      summary: "Returns the platform+locale specific data of the named Release"
      description: >
        Returns the platform+locale specific data of the named Release, which is a JSON Object. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#releases-release-builds-platform-locale)
      tags:
        - Releases
      operationId: auslib.web.common.releases.get_single_locale
      consumes: []
      produces:
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#releases-release-builds-platform-locale"
        description: Returns the platform+locale specific data of the named Release
      parameters:
        - $ref: '#/parameters/releaseParam'
        - $ref: '#/parameters/platformParam'
        - $ref: '#/parameters/localeParam'
      responses:
        '200':
          $ref: '#/responses/releaseDataResponse'
        '404':
          $ref: '#/responses/resourceNotFound'

    put:
      summary: Updates the named Release with the data given.
      description: >
        Updates the named Release with the data given. The 'blob' field is updated with the new one instead of being completely overridden. If the Release does not exist, it is created. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id11)
      tags:
        - Releases
      operationId: auslib.web.admin.views.releases.ensure_single_locale
      consumes:
        - application/json
      produces:
        - text/html
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id11"
        description: "Updates the named Release with the data given."
      parameters:
        - $ref: '#/parameters/releaseParam'
        - $ref: '#/parameters/platformParam'
        - $ref: '#/parameters/localeParam'
        - name: partial_release_body
          in: body
          description: Release object data
          required: true
          schema:
            $ref: '#/definitions/PartialRelease'
      responses:
        '200':
          $ref: '#/responses/modifiedDataVersion'
        '201':
          $ref: '#/responses/modifiedDataVersion'
        '400':
          $ref: '#/responses/invalidFormData'
        '401':
          $ref: '#/responses/unauthenticatedUser'
        '403':
          $ref: '#/responses/unauthorizedUser'
...
  /rules/{id_or_alias}:
    get:
      summary: Returns a rule identified by the id or alias
      description: >
        Returns the entire rule identified by the id or alias given in JSON format. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#rules-id-or-alias)
      tags:
        - Rules
      operationId: auslib.web.common.rules.get_rule
      consumes: []
      produces:
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id2"
        description: "Returns the entire rule identified by the id or alias given in JSON format."
      parameters:
        - $ref: '#/parameters/rule_id_or_alias_param'
      responses:
        '200':
          description: Successfully fetched a rule
          schema:
            $ref: "#/definitions/RuleObject"
          headers:
            X-Data-Version:
              description: data version of the rule
              type: integer
              format: int32
              minimum: 1
        '404':
          $ref: '#/responses/resourceNotFound'

    post:
      summary: Updates the rule
      description: >
        Updates the rule identified by the id or alias. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id3)
      tags:
        - Rules
      operationId: auslib.web.admin.views.rules.update_post
      consumes: []
      produces:
        - application/json
        - text/html
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id2"
        description: "Returns the entire rule identified by the id or alias given in JSON format."
      parameters:
        - $ref: '#/parameters/rule_id_or_alias_param'
        - name: rule
          in: body
          description: Rule object data
          required: true
          schema:
            allOf:
              - required:
                - data_version
              - $ref: '#/definitions/RulesBase'
              - $ref: '#/definitions/RulesNonNullable'
      responses:
        '200':
          $ref: '#/responses/updateExistingObject'
        '400':
          $ref: '#/responses/invalidFormData'
        '401':
          $ref: '#/responses/unauthenticatedUser'
        '403':
          $ref: '#/responses/unauthorizedUser'
        '404':
          $ref: '#/responses/resourceNotFound'

    # Exactly same as POST
    put:
      summary: Updates the rule
      description: >
        Updates the rule identified by the id or alias. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id3)
      tags:
        - Rules
      operationId: auslib.web.admin.views.rules.update_put
      consumes: []
      produces:
        - application/json
        - text/html
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#id2"
        description: "Returns the entire rule identified by the id or alias given in JSON format."
      parameters:
        - $ref: '#/parameters/rule_id_or_alias_param'
        - name: rule
          in: body
          description: Rule object data
          required: true
          schema:
            allOf:
              - required:
                - data_version
              - $ref: '#/definitions/RulesBase'
              - $ref: '#/definitions/RulesNonNullable'
      responses:
        '200':
          $ref: '#/responses/updateExistingObject'
        '400':
          $ref: '#/responses/invalidFormData'
        '401':
          $ref: '#/responses/unauthenticatedUser'
        '403':
          $ref: '#/responses/unauthorizedUser'
        '404':
          $ref: '#/responses/resourceNotFound'

    delete:
      summary: Deletes a rule identified by the rule id
      description: >
        Deletes the entire rule entry in DB identified by the rule_id. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#delete)
      tags:
        - Rules
      operationId: auslib.web.admin.views.rules.delete
      consumes: []
      produces:
        - application/json
        - text/html
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#delete"
        description: "Deletes the rule identified by the rule_id."
      parameters:
        - $ref: '#/parameters/rule_id_or_alias_param'
        - $ref: '#/parameters/dataVersionParam'
      responses:
        '200':
          description: Successfully deleted a rule
        '401':
          $ref: '#/responses/unauthenticatedUser'
        '403':
          $ref: '#/responses/unauthorizedUser'
        '404':
          $ref: '#/responses/resourceNotFound'
  /rules/{rule_id}/revisions:
    get:
      summary: Fetches all revisions of a Rule
      description: Fetches list and count of all the revisions of the rule given by rule_id
      tags:
        - Rules
      operationId: auslib.web.common.rules.get_history
      consumes: []
      produces:
        - application/json
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/database.html#rules"
        description: "Fetches list and count of all the revisions of the rule given by rule_id"
      parameters:
        - $ref: '#/parameters/rule_id_param'
        - $ref: '#/parameters/pageParam'
        - $ref: '#/parameters/limitParam'
      responses:
        '200':
          description: Returns JSON object of list of rules with their count
          schema:
            type: object
            allOf:
              - $ref: "#/definitions/RulesListGET"
              - type: object
                properties:
                  rules:
                    description: array where each element is a rule object plus timestamp, change_id and changed_by
                    type: array
                    items:
                      $ref: '#/definitions/HistoryModel'
          examples:
            application/json:
              count: 1
              rules:
                - alias: "esr45-localtest"
                  backgroundRate: 100
                  buildID: null
                  buildTarget: null
                  channel: "esr-localtest"
                  comment: "esr45 esr-localtest rule"
                  changed_by: "balrogadmin"
                  change_id: 10
                  timestamp: 149530544
                  data_version: 22
                  distVersion: null
                  distribution: null
                  fallbackMapping: null
                  headerArchitecture: null
                  instructionSet: null
                  locale: null
                  mapping: "Firefox-45.7.0esr-build1"
                  memory: null
                  mig64: null
                  osVersion: null
                  priority: 43
                  product: "Firefox"
                  rule_id: 308
                  update_type: "minor"
                  version: null
                  allowlist: null
        '404':
          $ref: '#/responses/resourceNotFound'
        '400':
          $ref: '#/responses/invalidFormData'

    post:
      summary: Reverts the rule to a previous revision
      description: >
        Updates the rule identified by the rule_id to a previous revision identified by change_id. [Docs](http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#history-view-object-change-id-field)
      tags:
        - Rules
      operationId: auslib.web.admin.views.rules.revert_to_older_revision
      consumes:
        - application/json
      produces:
        - application/json
        - text/html
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#history-view-object-change-id-field"
        description: "Updates the rule identified by the rule_id to a previous revision identified by change_id."
      parameters:
        - $ref: '#/parameters/rule_id_param'
        - $ref: '#/parameters/change_id_body_param'
      responses:
        '200':
          description: Successfully reverted back the Rule
        '400':
          $ref: '#/responses/invalidFormData'
        '401':
          $ref: '#/responses/unauthenticatedUser'
        '403':
          $ref: '#/responses/unauthorizedUser'
        '404':
          $ref: '#/responses/resourceNotFound'
...
definitions:
...
  ReleaseData:
    title: Release form values plus blob
    description: "Release Base Model + release blob"
    type: object
    allOf:
      - required:
        - name
        - product
        - blob
      - $ref: '#/definitions/DataVersionModel'
      - $ref: '#/definitions/ProductModel'
      - $ref: '#/definitions/ReleaseName'
      - type: object
        properties:
          blob:
            type: string
            example: '{"hashFunction": "sha512", "name": "CDM-979", "product": "Firefox", "schema_version": 1000, "vendors": { "gmp-eme-adobe": { "platforms": {"WINNT_x86_64-msvc-x64": {"alias": "WINNT_x86_64-msvc"}},"version": "15"}}}'
            format: JSONStringField
            description: json blob containing information about the release
...
  ReleaseDataField:
    title: Release Data field
    description: "release's json blob data as marshalled as string"
    type: object
    properties:
      data:
        description: json blob containing information about the release marshalled as string
        type: string
        format: JSONStringField
        example: '{"hashFunction": "sha512", "name": "CDM-979", "product": "Firefox", "schema_version": 1000, "vendors": { "gmp-eme-adobe": { "platforms": {"WINNT_x86_64-msvc-x64": {"alias": "WINNT_x86_64-msvc"}},"version": "15"}}}'
  PartialRelease:
    title: Partial Release
    description: "Partial Release form values"
    type: object
    allOf:
      - required:
        - product
        - data
      # Note: 'name' is inherited with other required fields from Base model but never used when processing the input request
      - $ref: '#/definitions/DataVersionModel'
      - $ref: '#/definitions/ProductModel'
      - $ref: '#/definitions/ReleaseName'
      - $ref: '#/definitions/ReleaseDataField'
      - type: object
        properties:
          hashFunction:
            description: cryptographic hash function to assure integrity of transmitted data
            type: string
            format: ascii
            example: sha512
          schema_version:
            description: The version number of release
            type: integer
            format: int32
            minimum: 1
            example: 1
          copyTo:
            description: List names of releases which the updated data must be copied to
            type: string
            format: ascii
            format: JSONStringField
          alias:
            description: id of the release
            type: string
            format: ascii
            format: JSONStringField
...